feat: Route protection + Minor bug fix with Logout

+++ frontend/src/components/PostDetails.jsx
@@ -1,11 +1,20 @@
 import { usePostsContext } from "../hooks/usePostsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 const PostDetails = ({post})=>{
     const {dispatch} = usePostsContext()
+    const {user} = useAuthContext()
     
     const handleClick = async()=>{
+        if(!user){
+            return
+        }
+
         const response = await fetch('http://localhost:4000/api/posts/' + post._id,{
-            method : 'DELETE'
+            method : 'DELETE',
+            headers : {
+                'Authorization' : `Bearer ${user.token}`
+            }
         })
 
         const json =  await response.json()
